feat: rename error utility file
