+++ packages/graphql-yoga/src/error.ts
